+++ pyvisim/features/__init__.py
@@ -1,3 +1,15 @@
-from ._features import SIFT, DeepConvFeature, Lambda, RootSIFT
+from ._features import (
+    SIFT,
+    DeepConvFeature,
+    Lambda,
+    RootSIFT,
+    feature_extractor_from_dict,
+)
 
-__all__ = ["SIFT", "RootSIFT", "DeepConvFeature", "Lambda"]
+__all__ = [
+    "SIFT",
+    "RootSIFT",
+    "DeepConvFeature",
+    "Lambda",
+    "feature_extractor_from_dict",
+]